requirements.txt
openai
pandas
numpy
Flask
flask-cors
python-dotenv
gunicorn